Revert "Bump postcss from 8.4.20 to 8.4.31 in /merged-packages/eth-json-rpc-provider" (#1790)
## Explanation

- The contents of `merged-packages/` shouldn't be altered, as it's a
temporary directory used for migrating packages into the monorepo.
- Any dependency updates for the migration target package need to be
handled manually. The migration process involves aligning dependency
versions with the monorepo (which may not use the latest version),
removing dependencies that are defined in the monorepo root, and
removing the `yarn.lock` file from the migration target subpackage.

## Reference

- Blocks https://github.com/MetaMask/core/pull/1764
- See
https://github.com/MetaMask/core/issues/1079#issuecomment-1700126302 for
context on the role of `merged-packages/`.