Remove `for..in` loops and reenable eslint rules: `guard-for-in`, `no-for-in-array` (#1865)
## Explanation

`for..in` loops [iterate over the entire prototype
chain](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Statements/for...in)
and are discouraged because they can cause unexpected behavior. Our
ESLint ruleset also guards against them by default.

`for..in` loops are sometimes used because they are useful for
introspection of object properties, but there are safer alternatives:

- Use `for..of Object.keys()` if the index signature is just `string`.
- When inspecting property keys with signatures that are more specific
than `string`, use `getKnownPropertyNames()` from `@metamask/utils` as a
safe alternative to `for..in`.
- Use `for..of Object.keys(obj) as (keyof typeof obj)[])` for stronger
typing of the index signature if `obj` properties are generic, populated
at runtime, or otherwise not fully determinable by
`getKnownPropertyNames()`.
- To access array indices from a `for..of` loop, use the
`Array.prototype.entries()` method.

This PR removes all `for..in` loops in the core monorepo, and resolves
all conflicts with the ESLint rules `guard-for-in` and
`@typescript-eslint/no-for-in-array`. Some typing improvements are also
introduced.

## References

- Closes #1866 

## Changelog

No consumer-facing changes other than the reenabled linter rules.

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>