Remove files and directories that will be replaced by files in the mo… (#1875)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
In the process of migrating json-rpc-engine package into core monorepo,
in this PR, we are remove files and directories that will be replaced by
files in the monorepo root.
Remove: .github/, .git*, scripts/, .depcheckrc.json, .yarn/,
.yarnrc.yml, yarn.lock, .editorconfig, .eslint*, .prettier*, .nvm*.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->
- See #1551 

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate