json-rpc-engine: Update tsconfig*.json, typedoc.json, jest.config.js to align with repo defaults (#1878)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
In the process of migrating json-rpc-engine to core monorepo. In this
PR, we have updated tsconfig*.json, typedoc.json, jest.config.js to
extend from corresponding files in the core root directory.


## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Jongsun Suh <34228073+MajorLift@users.noreply.github.com>