Update `PollingController` to immediately call `_executePoll` if no polling is currently active for the key on start (#1874)
## Explanation

Currently when PollingController polling is started via
`startPollingByNetworkClientId()`, the first poll execution doesn't
occur until `intervalLength` time has passed. Some controllers such as
the `CurrencyRateController` want to execute their polling logic
immediately on polling start that way it is available to the UI as soon
as possible. ~~This PR adds a `executeImmediately` flag to the
`PollingController`. When set to `true`, it causes `_executePoll()` to
immediately be called when `startPollingByNetworkClientId()` is called,
subsequently on each `intervalLength` time period. This value defaults
to false.~~ This PR updates `#poll()` to call `_executePoll()` if no
polling is currently active for the key

## References

* Related https://github.com/MetaMask/core/pull/1805

## Changelog

### `@metamask/polling-controller`

- **BREAKING**: `_executePoll()` is called immediately on start if no
polling interval is already active for the networkClientId + options
combination

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Elliot Winkler <elliot.winkler@gmail.com>
Co-authored-by: Alex <adonesky@gmail.com>