Fix: Token detection on accounts other than current (#1848)
## Explanation

`TokensController.addDetectedTokens` supports detecting tokens on
accounts other than the one currently selected. But a bug occurs in this
case. Tokens (both added and detected) from the current account get
incorrectly copied to the detecting account, even when the detecting
account has no balance of those tokens.

This happened because `state.tokens` and `state.detectedTokens` were
used to check for existing tokens, but these are scoped to the current
account rather than the detecting account.

Fix: Check `state.allTokens`, `state.allDetectedTokens`, and
`state.allIgnoredTokens` instead, since they contain tokens for all
chains + addresses. Some perf improvements are also made to combine
`find` and `indexOf` into a single `findIndex`.

Added a unit test that reproduced the problem on the old code, and
passes with the new code. The existing test did not catch this because
of its operation order (it detected the other account first, then added
to the current account).


## Changelog

### `@metamask/assets-controllers`

- **FIXED**: Token detection on accounts other than current

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate