Typing fixes for `TransactionController.test.ts` (#1954)
With the Transaction Controller under active development, it's important to ensure that its tests are valid. This PR improves typing in the tests by removing `any` usage and fixing `as` assertions, mainly by enforcing that `TransactionMeta` mock objects correctly conform to the type they're intended to test. 

Typing mock objects as `any` can be convenient for omitting required but non-essential properties, but it also keeps TypeScript from alerting us when objects with invalid shapes are being used for tests. If necessary, type assertions that don't use `any` e.g. `as TransactionMeta` are a much better compromise.

- Fixed test mock object typing:
  - `TransactionMeta` with failed status should contain an error property.
  - `from`, `to` should be located in the `txParams` property (or the mock object typed as `TransactionMeta & { from: Hex; to: Hex }`)
- Removed `any` usage that's become unnecessary (a key problem with `any` is that it doesn't tell you when it's no longer needed, and continues to suppress any changed or new errors).
- Changed `TransactionMeta` definition to equivalent but more readable expression.