Use `PollingController` in `CurrencyRateController` and key by nativeCurrency (#1805)
## Explanation

Currently the CurrencyRateController only supports a single
nativeCurrency (ticker) to be polled at a time. Additionally, it uses an
odd notion of `pendingNativeCurrency` and `pendingCurrentCurrency` which
are used as intermediary flags for the goal of ultimately fetching the
new conversion rate for the NativeCurrency (fiat) to CurrentCurrency
(ticker) pair.

We need to add support for multiple tickers being polled concurrently.
This is achieved by using the PollingController and changing the
conversion rates to be keyed by ticker. Additionally, changing the
currentCurrency now clears the existing conversion rate state until it
is refetched (which gets rid of the need for pendingCurrentCurrency).

 
## References

See https://github.com/MetaMask/MetaMask-planning/issues/1025

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **BREAKING**: `CurrencyRateController` is now keyed by
`nativeCurrency` (i.e. ticker) for `conversionDate`, `conversionRate`,
and `usdConversionRate` in the `currencyRates` object. `nativeCurrency`,
`pendingNativeCurrency`, and `pendingCurrentCurrency` have been removed.
  - ```
    export type CurrencyRateState = {
      currentCurrency: string;
      currencyRates: Record<
        string,
        {
          conversionDate: number | null;
          conversionRate: number | null;
          usdConversionRate: number | null;
        }
      >;
    };
    ```
- **BREAKING**: `CurrencyRateController` now extends `PollingController`
- `start()` and `stop()` methods replaced with
`startPollingByNetworkClientId()`, `stopPollingByPollingToken()`, and
`stopAllPolling()`
- **BREAKING**: `CurrencyRateController` now sends the
`NetworkController:getNetworkClientById` action via messaging controller


## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Alex Donesky <adonesky@gmail.com>