Bump npm-publish action to 4.x
Currently, when publishing a new release of the monorepo, an attempt is
made to publish packages whose versions have not changed and therefore
do not need publishing. This causes the `publish-npm` step in the
release workflow to appear to fail, which can be seen by looking at any
recent release of the monorepo ([this one][1], for instance).

This problem is solved in the current version of `action-npm-publish`,
but we are not using this version consistently: we are using it for the
dry run, but not for the "real" run.

This commit bumps this action to 4.x so that this problem goes away.
I've confirmed this on a [fork of `core`][2] (this run also failed, but
only because the NPM token was invalid; you should be able to see that
it skipped all packages but the first).

[1]: https://github.com/MetaMask/core/actions/runs/6712993797/job/18243918833
[2]: https://github.com/mcmire/core/actions/runs/6749847699/job/18351193396