Add `onBootCleanup` functionality to constructor and `onNetworkChange` (#1916)
## Explanation

This PR aims to bring `onBootCleanup` functionality to the core
`TransactionController`. It is also removing the `initApprovals`
function because it's unnecessary to do it in client, and put that logic
into `onBootCleanup`.

PR is not going to be marked as _breaking change_ since `initApprovals`
is not in the use of mobile anyways.

### `loadingDefaults` property removed

This property is been used to tell the ui when we are done calculating
gas but it’s not used anymore (in the extension, nor in mobile), so
while bringing this functionality to core, I decided to delete it.

## References

Fixes https://github.com/MetaMask/MetaMask-planning/issues/1473

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/transaction-controller`

- **Added**: `onBootCleanup` functionality to constructor and
`onNetworkChange`
- Initialises approvals of `TransactionStatus.unapproved` transactions
on active network
- Refreshes the gas values for `TransactionStatus.unapproved`
transactions on active network
- Submits the `TransactionStatus.approved` transactions on active
network

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Matthew Walsh <matthew.walsh@consensys.net>