Add hooks to support mmi logic (#1787)
## Explanation

**Context**
As part of the effort to unify the transaction controller between
extensions and core a crucial part is replacing the MMI code fences
using callbacks and arguments in the extension
[PR](https://github.com/MetaMask/metamask-extension/pull/21171) as the
core `TransactionController` is a static package and cannot have any
build specific code.

This PR aims to add/adopt suitable callbacks and arguments to support
current MMI logic based on the PR above. It adds optional `hooks` in the
constructions and includes callbacks such as `afterSign`,
`beforeApproveOnInit`, `beforePublish` and `getAdditionalSignArguments`.
These hooks give the possibility to the client execute additional logic
in between flows.

- **ADDED**: add the option in the construction `hooks` giving the
possibility to pass callback functions additional logic to execute.
- **ADDED**: add the optional callback in hooks
`getAdditionalSignArguments` returns additional arguments required to
sign a transaction.
- **ADDED**: add the optional callback in hooks `beforeApproveOnInit`
additional logic to execute before starting an approval flow for a
transaction during initialization. Return false to skip the transaction.
**ADDED**: add the optional callback in hooks `afterSign` additional
logic to execute after signing a transaction. Return false to not change
the status to signed.
**ADDED**: add the optional callback in hooks `beforePublish` additional
logic to execute before publishing a transaction. Return false to
prevent the broadcast of the transaction.