Align NetworkController provider type w/ eth-query (#2028)
Recent changes in `@metamask/utils` caused SafeEventEmitterProvider, the
type of the provider that NetworkController exposes, to no longer align
with the provider that `@metamask/eth-query` expects. We temporarily
worked around this with the use of `// @ts-expect-error`, but this
merely hid the problem. A fix was made in `@metamask/eth-query`, so this
commit bumps that package and removes the associated `//
@ts-expect-error` lines.