Extract/fix FakeProvider for use among packages (#2039)
The FakeProvider class is useful in tests because it allows you to pass
a provider that doesn't make HTTP requests to a library such as
`eth-query`.

Previously, this class was only available in NetworkController tests. We
want to now start using it in other tests, specifically along with
`eth-query`. However, one problem I noticed is that it doesn't conform
to SafeEventEmitterProvider completely. So this commit not only moves it
to a more global location but also corrects its API so that it can be
passed to `eth-query`.