refactor(base-controller): Replace `private` keyword
All uses of the `private` keyword have been replaced with a `#` prefix
(except those in the deprecated `BaseController`). The use of `private`
is discouraged in our repositories, as `#`-prefixed properties are
truly private and are easier to work with.

Technically this could be considered a breaking change because there
are ways to work around `private`, but I would consider this non-
breaking because these properties were clearly intended to be private.