Add getNetworkConfigurationByNetworkClientId
Currently, if you have the ID of a network client for a custom network
and you want to get information like the ticker or nickname for that
network, you have to have a reference to the network controller so you
can look up that network client ID in the list of network
configurations. It's very roundabout, and we can do better.

This commit introduces a method on NetworkController,
`getNetworkConfigurationByNetworkClientId`, that can be used to retrieve
the network configuration that corresponds to a custom network given a
network client ID. Although the concept of a network configuration was
created for custom networks, this method also works for a built-in
network client ID by constructing a "fake" network configuration.

This commit also introduces a controller action so that the network
controller instance is not needed in order to call this method.