feat(base-controller): Allow using internal events/actions (#2050)
## Explanation

The restricted controller messenger now allows using all internal events
and actions. Previously internal events were treated the same as
external events, they had to be explicitly allowed.

This is a non-breaking change; internal actions/events are still
permitted to be list as "allowed", it's just no longer necessary.

## References

Resolves #2047

## Changelog

### `@metamask/base-controller`

- CHANGED: The restricted controller messenger now allows calling all
internal events and actions
- Previously internal events and actions were only usable if they were
listed as "allowed". They are still permitted to be listed as "allowed"
now, but it is no longer necessary.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Co-authored-by: Jongsun Suh <jongsun.suh@icloud.com>