Support token detection on Arbitrum and Optimism (#2035)
## Explanation

Adds support for token detection on arbitrum and optimism. The goal is
to eventually support token detection on MetaMask's top 10 most popular
networks.

Token lists are available for both chains:
- https://token-api.metaswap.codefi.network/tokens/42161
- https://token-api.metaswap.codefi.network/tokens/10

As well as the ERC20 balance checker:
- https://github.com/wbobeirne/eth-balance-checker#deployed-addresses

Related PR to extension (depends on this PR):
https://github.com/MetaMask/metamask-extension/pull/21841

## References

Related issues:
- https://github.com/MetaMask/metamask-extension/issues/17697
- https://github.com/MetaMask/metamask-extension/issues/16496

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **ADDED**: Token detection support for Arbitrum and Optimism

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate