Get token prices from our API instead of CoinGecko
CoinGecko is planning on retiring their `/simple/token_price` endpoint
soon. This endpoint is critical for us, as we use to fetch prices for
tokens. This commit removes all code related to CoinGecko in
TokenRatesController so that it hits an internal API instead.

This is set up in such a way that we can upgrade to newer versions of
the API (or even other services) in the future without breaking backward
compatibility. However, at the moment it does introduce breaking
changes to TokenRatesController.