TokenListController: Get prefiltered tokens (#2054)
The list of tokens that the token list controller fetches is 2.2 MB.
After fetching, the controller filters out duplicate and rare tokens.
Making this request and processing this data is increasing app load
time.

The filtering logic that the controller performs has now been moved to
the API, so this commit instructs the API to do this by updating the
token list API URL with a set of query parameters. This means that the
API will return a prefiltered list of tokens, alleviating the
performance issues.