fix(assets-controllers): TokenRatesController state consistency
The TokenRatesController has been updated to ensure that the two sets
of exchange rate state are always consistent with each other.
Previously if the method `updateExchangeRatesByChainId` was called
directly, it would not update the `contractExchangeRates` state even if
the equivalent part of `contractExchangeRatesByChainId` was updated by
that call.

The tests have undergone a substantial refactor to ensure that we cover
all cases between both of these update methods. The test cases written
previously left out many test cases. We should now have better (but not
perfect) coverage, and the tests are now shared between the two methods
because they have identical test cases.

Fixes #3597